Fixed the wake sleep issue on a reload.  The times were not being serialized out.

+++ Cluster_Search/SerializeHelper.cs
@@ -12,6 +12,11 @@ namespace SEARCH
    {
      public static void SerializeObjectToFile(string inFileName, object inO)
       {
+         string backupDir = Path.GetDirectoryName(inFileName);
+        if (! Directory.Exists(backupDir))
+        {
+         Directory.CreateDirectory(backupDir);
+         }
          StreamWriter sw = new StreamWriter(inFileName);
          XmlSerializer ser = new XmlSerializer(inO.GetType());
          ser.Serialize(sw, inO);
